Added numba optimisation
Added numba optimisation to the _upwind_step_ routine, but not to the time loops (which are currently embedded within the HUXt1D and HUXt2D classes. Future change - have HUXt classes generate the input time series, which is then solved by external optimised function

Generalised HUXt1d and HUXt2D classes to allow the inner and outer boundary positions to be user defined

Added a fast solver for timeseries at a fixed position, for use in large CME ensembles. Updated jupyter labbook with an example

Updated environment to include numba